backend/requirements.txt
Flask
flask.cors
flask_sqlalchemy
flask_jwt_extended